# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/webhook-endpoints": {
      "post": {
        "description": "Returns the signing secret (`whsec_…`) once; store it safely. Test-mode endpoints receive sandbox events, live endpoints live events. The URL must use HTTPS and reach the public internet.",
        "operationId": "WebhookEndpointsController_create",
        "parameters": [
          {
            "name": "BitoCard-Mode",
            "in": "header",
            "description": "Dashboard sessions only: test for the sandbox, live (default) otherwise. API keys always use their own mode.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEndpointDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a webhook endpoint",
        "tags": [
          "Webhooks"
        ]
      },
      "get": {
        "operationId": "WebhookEndpointsController_list",
        "parameters": [
          {
            "name": "BitoCard-Mode",
            "in": "header",
            "description": "Dashboard sessions only: test for the sandbox, live (default) otherwise. API keys always use their own mode.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List webhook endpoints",
        "tags": [
          "Webhooks"
        ]
      }
    },
    "/v1/webhook-endpoints/{id}": {
      "get": {
        "operationId": "WebhookEndpointsController_get",
        "parameters": [
          {
            "name": "BitoCard-Mode",
            "in": "header",
            "description": "Dashboard sessions only: test for the sandbox, live (default) otherwise. API keys always use their own mode.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a webhook endpoint",
        "tags": [
          "Webhooks"
        ]
      },
      "patch": {
        "description": "Change its URL, description or event types, or enable it again after it was disabled.",
        "operationId": "WebhookEndpointsController_update",
        "parameters": [
          {
            "name": "BitoCard-Mode",
            "in": "header",
            "description": "Dashboard sessions only: test for the sandbox, live (default) otherwise. API keys always use their own mode.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEndpointDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a webhook endpoint",
        "tags": [
          "Webhooks"
        ]
      },
      "delete": {
        "description": "Its delivery log is deleted too. Events stay available from `GET /v1/events`.",
        "operationId": "WebhookEndpointsController_remove",
        "parameters": [
          {
            "name": "BitoCard-Mode",
            "in": "header",
            "description": "Dashboard sessions only: test for the sandbox, live (default) otherwise. API keys always use their own mode.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a webhook endpoint",
        "tags": [
          "Webhooks"
        ]
      }
    },
    "/v1/webhook-endpoints/{id}/rotate-secret": {
      "post": {
        "description": "Returns the new secret once. Until the old one expires, each delivery carries two `v1` signatures, one for each secret.",
        "operationId": "WebhookEndpointsController_rotate",
        "parameters": [
          {
            "name": "BitoCard-Mode",
            "in": "header",
            "description": "Dashboard sessions only: test for the sandbox, live (default) otherwise. API keys always use their own mode.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RotateSecretDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rotate the signing secret",
        "tags": [
          "Webhooks"
        ]
      }
    },
    "/v1/webhook-endpoints/{id}/test": {
      "post": {
        "description": "Sends a `ping` event to this endpoint now and returns the delivery with its result. Test events are not retried.",
        "operationId": "WebhookEndpointsController_test",
        "parameters": [
          {
            "name": "BitoCard-Mode",
            "in": "header",
            "description": "Dashboard sessions only: test for the sandbox, live (default) otherwise. API keys always use their own mode.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send a test event",
        "tags": [
          "Webhooks"
        ]
      }
    },
...
    "/v1/events": {
      "get": {
        "description": "Every event from the last 30 days, oldest first. Pass the ID of the last event you handled as `since` to catch up on anything missed; keep calling while `has_more` is true. Events appear here a few seconds after they happen. This list, not the webhook, is the source of truth.",
        "operationId": "EventsController_list",
        "parameters": [
          {
            "name": "BitoCard-Mode",
            "in": "header",
            "description": "Dashboard sessions only: test for the sandbox, live (default) otherwise. API keys always use their own mode.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "required": false,
            "in": "query",
            "description": "The ID of the last event you handled, or an ISO 8601 time. Events after it are returned, oldest first.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "order.completed",
                "order.failed",
                "order.refunded",
                "top_up.succeeded",
                "top_up.failed",
                "payout.paid",
                "payout.failed"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List events",
        "tags": [
          "Events"
        ]
      }
    },
    "/v1/events/{id}": {
      "get": {
        "operationId": "EventsController_get",
        "parameters": [
          {
            "name": "BitoCard-Mode",
            "in": "header",
            "description": "Dashboard sessions only: test for the sandbox, live (default) otherwise. API keys always use their own mode.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get an event",
        "tags": [
          "Events"
        ]
      }
    },
...
  },
  "info": {
    "title": "BitoCard API",
    "description": "One API for catalogue, quotes, orders, wallets and webhooks. Sandbox and live share this address; your API key decides the mode.",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.bitocard.com"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "API key",
        "type": "http",
        "description": "API key: bc_test_… (sandbox) or bc_live_… (live)."
      }
    },
    "schemas": {
...
      "CreateEndpointDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://example.com/webhooks/bitocard",
            "description": "An HTTPS address on the public internet."
          },
          "description": {
            "type": "string",
            "maxLength": 200
          },
          "events": {
            "type": "array",
            "default": [
              "*"
            ],
            "description": "Event types to receive; `*` for all, including types added later.",
            "items": {
              "type": "string",
              "enum": [
                "*",
                "order.completed",
                "order.failed",
                "order.refunded",
                "top_up.succeeded",
                "top_up.failed",
                "payout.paid",
                "payout.failed"
              ]
            }
          }
        },
        "required": [
          "url"
        ]
      },
      "UpdateEndpointDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "maxLength": 200
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "*",
                "order.completed",
                "order.failed",
                "order.refunded",
                "top_up.succeeded",
                "top_up.failed",
                "payout.paid",
                "payout.failed"
              ]
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "enabled",
              "disabled"
            ],
            "description": "Enable an endpoint again after fixing it, or pause it."
          }
        }
      },
      "RotateSecretDto": {
        "type": "object",
        "properties": {
          "expire_previous_in_hours": {
            "type": "number",
            "minimum": 0,
            "maximum": 168,
            "default": 24,
            "description": "Hours the old secret keeps signing alongside the new one; 0 stops it now."
          }
        }
      },
...
    }
  },
...
}